Refactored Kconfig for USB subsystem.
Definitions of UsbDevice moved to source files.
Removed platform-specific usb_device.h headers.
Added separate EP queue size and descriptor pool size options.

+++ platform/nxp/lpc13xx/usb_base.c
@@ -55,7 +55,7 @@ static void configPins(struct UsbBase *device,
       config->dm, config->dp, config->connect, config->vbus
   };
 
-  for (unsigned int index = 0; index < ARRAY_SIZE(pinArray); ++index)
+  for (size_t index = 0; index < ARRAY_SIZE(pinArray); ++index)
   {
     if (pinArray[index])
     {
